requirements.txt
openai-whisper
sounddevice
numpy
torch
pyperclip
